+++ src/icons/SearchIcon.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+
+export default function SearchIcon() {
+  return (
+    <svg
+      width="18"
+      height="18"
+      viewBox="0 0 18 18"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M8.0625 0.875C8.0625 0.875 9.52438 0.875 10.8604 1.44009C10.8604 1.44009 12.1504 1.98569 13.1448 2.98017C13.1448 2.98017 14.1393 3.97465 14.6849 5.26461C14.6849 5.26461 15.25 6.60062 15.25 8.0625C15.25 8.0625 15.25 9.52438 14.6849 10.8604C14.6849 10.8604 14.1393 12.1504 13.1448 13.1448C13.1448 13.1448 12.1504 14.1393 10.8604 14.6849C10.8604 14.6849 9.52438 15.25 8.0625 15.25C8.0625 15.25 6.60062 15.25 5.26461 14.6849C5.26461 14.6849 3.97465 14.1393 2.98017 13.1448C2.98017 13.1448 1.98569 12.1504 1.44009 10.8604C1.44009 10.8604 0.875 9.52438 0.875 8.0625C0.875 8.0625 0.875 6.60062 1.44009 5.26461C1.44009 5.26461 1.98569 3.97465 2.98017 2.98017C2.98017 2.98017 3.97464 1.9857 5.26461 1.44009C5.26461 1.44009 6.60062 0.875 8.0625 0.875ZM8.0625 2.125C8.0625 2.125 6.8541 2.125 5.75155 2.59134C5.75155 2.59134 4.68613 3.04198 3.86405 3.86405C3.86405 3.86405 3.04197 4.68613 2.59134 5.75155C2.59134 5.75155 2.125 6.8541 2.125 8.0625C2.125 8.0625 2.125 9.2709 2.59134 10.3735C2.59134 10.3735 3.04197 11.4389 3.86405 12.2609C3.86405 12.2609 4.68613 13.083 5.75155 13.5337C5.75155 13.5337 6.85411 14 8.0625 14C8.0625 14 9.27089 14 10.3735 13.5337C10.3735 13.5337 11.4389 13.083 12.2609 12.2609C12.2609 12.2609 13.083 11.4389 13.5337 10.3735C13.5337 10.3735 14 9.27089 14 8.0625C14 8.0625 14 6.85411 13.5337 5.75155C13.5337 5.75155 13.083 4.68613 12.2609 3.86405C12.2609 3.86405 11.4389 3.04197 10.3735 2.59134C10.3735 2.59134 9.2709 2.125 8.0625 2.125Z"
+        fill="#9D9E9E"
+      />
+      <path
+        d="M16.0578 16.9416C16.175 17.0589 16.3343 17.125 16.5 17.125C16.6658 17.125 16.8248 17.0592 16.942 16.942C17.0592 16.8248 17.125 16.6658 17.125 16.5C17.125 16.3343 17.0592 16.1753 16.942 16.0581L13.1452 12.2613C13.0279 12.144 12.8689 12.0782 12.7032 12.0782C12.5374 12.0782 12.3782 12.1437 12.2609 12.2609C12.1437 12.3782 12.0782 12.5374 12.0782 12.7032C12.0782 12.8689 12.144 13.0279 12.2612 13.1451L16.0578 16.9416Z"
+        fill="#9D9E9E"
+      />
+    </svg>
+  );
+}
