+++ src/pages/_app.tsx
@@ -4,12 +4,14 @@ import { useEffect, useState } from "react";
 import { getUser } from "../../common/getUser";
 import { GetUserInfoResponse } from "@/model/users/users";
 import Script from "next/script";
+import style from "../styles/app.module.css";
 
 function Loading() {
   return (
     <div className="flex items-center justify-center w-full h-full z-40 fixed">
-      <span className="animate-ping absolute inline-flex h-10 w-10 rounded-full bg-textWhite opacity-75" />
-      <span className="inline-flex rounded-full h-7 w-7 bg-sky-500" />
+      {/* <span className="animate-ping absolute inline-flex h-10 w-10 rounded-full bg-textWhite opacity-75" />
+      <span className="inline-flex rounded-full h-7 w-7 bg-sky-500" /> */}
+      <div className={style.loader}>Loading...</div>
     </div>
   );
 }
